+++ geotrek/trekking/static/trekking/infrastructurelayer.js
@@ -16,7 +16,7 @@ var InfrastructuresLayer = L.GeoJSON.extend({
         }
 
         var img = L.Util.template('<img src="{SRC}" title="{TITLE}">', {
-            SRC: featureData.properties.type.pictogram || '/static/trekking/infrastructure.png',
+            SRC: featureData.properties.type.pictogram,
             TITLE: infrastructureLabel
         });