+++ docs/changelog.rst
@@ -2,6 +2,29 @@
 CHANGELOG
 =========
 
+2.22.3 (2018-12-14)
+-------------------
+
+**Bug fixes**
+
+- Don't publish deleted infrastructures/signages
+- Add default pictograms to published infrastructures/signages
+
+
+2.22.2 (2018-12-10)
+-------------------
+
+**Bug fixes**
+
+- Fix bugs with HTTPS access
+- Fix for some modules to edit attributes and not the geometry
+
+**Minor changes**
+
+- add options to sync signages and infrastructures
+- sync global signages and infrastructures
+
+
 2.22.1 (2018-11-27)
 -------------------